# contracts/openapi/tab-v1.yaml
openapi: 3.0.3
info:
  title: UTab API v1
  version: 0.1.1
  description: >
    Contract-first API used by tournament operations. This is the initial
    machine-readable contract for Phase 0.
servers:
  - url: https://tab.example.com
security:
  - sessionAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v1/speakers/{id}/personal-data:
    delete:
      tags: [Privacy]
      summary: Erase speaker personal data for one tournament
      operationId: eraseSpeakerPersonalData
      parameters:
        - $ref: "#/components/parameters/SpeakerIdPath"
        - $ref: "#/components/parameters/TournamentIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErasureRequestBody"
      responses:
        "200":
          description: Erased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeErasureResult"
...
  /api/v1/adjudicators/{id}/personal-data:
    delete:
      tags: [Privacy]
      summary: Erase adjudicator personal data for one tournament
      operationId: eraseAdjudicatorPersonalData
      parameters:
        - $ref: "#/components/parameters/AdjudicatorIdPath"
        - $ref: "#/components/parameters/TournamentIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErasureRequestBody"
      responses:
        "200":
          description: Erased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeErasureResult"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SpeakerIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Speaker ObjectId
    AdjudicatorIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Adjudicator ObjectId
    TournamentIdQuery:
      name: tournamentId
      in: query
      required: true
      schema:
        type: string
      description: Tournament ObjectId
  schemas:
...
    EnvelopeErasureResult:
      type: object
      required: [data, errors]
      properties:
        data:
          $ref: "#/components/schemas/ErasureResult"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"
...
    ErasureResult:
      type: object
      required: [tournamentId, entityType, entityId, redacted]
      properties:
        tournamentId:
          type: string
        entityType:
          type: string
          enum: [speaker, adjudicator]
        entityId:
          type: string
        redacted:
          type: boolean
        eraseMode:
          type: string
          enum: [anonymize, hard_delete]
        reason:
          type: string
        approvedBy:
          type: string
          nullable: true
        targetRefs:
          type: array
          items:
            type: string
        submissionCommentsCleared:
          type: integer
          minimum: 0
...
    ErasureRequestBody:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 5
          maxLength: 500
        approvedBy:
          type: string
          minLength: 1
          maxLength: 128
        targetRefs:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 256
        eraseMode:
          type: string
          enum: [anonymize, hard_delete]
        reauthPassword:
          type: string
          minLength: 1
          maxLength: 200
          description: >
            Required for interactive organizer flows. Service-account callers
            are exempt.
...
    ErrorItem:
      type: object
      required: [name, message]
      properties:
        name:
          type: string
        message:
          type: string